+++ college_flow_app/lib/config/routes/flow_router.dart
@@ -1,27 +1,28 @@
 import 'package:college_flow_app/config/routes/flow_routes.dart';
 import 'package:college_flow_app/presentation/login/login_pages.dart';
-import 'package:college_flow_app/presentation/login/register_geral.dart';
-import 'package:college_flow_app/presentation/login/register_page.dart';
-import 'package:college_flow_app/presentation/login/register_student.dart';
 import 'package:college_flow_app/presentation/walkthrough/walkthrough_page.dart';
 import 'package:college_flow_app/utils/helpers/navigator_helper.dart';
 import 'package:flutter/material.dart' hide Router;
+
+import '../../presentation/register/register_geral.dart';
+import '../../presentation/register/register_page.dart';
+import '../../presentation/register/register_student.dart';
 
 abstract class FlowRouter {
   static Route routeFromSettings(RouteSettings settings) {
     late Widget page;
 
     switch (settings.name) {
-      case FlowRoutes.RegisterGeral:
+      case FlowRoutes.registerGeral:
         page = const RegisterGeral();
         break;
-      case FlowRoutes.RegisterStudent:
+      case FlowRoutes.registerStudent:
         page = const RegisterStudent();
         break;
-      case FlowRoutes.RegisterPage:
+      case FlowRoutes.registerPage:
         page = const RegisterPage();
         break;
-      case FlowRoutes.LoginPage:
+      case FlowRoutes.loginPage:
         page = const LoginPage();
         break;
       case FlowRoutes.mainPage:
