+++ college_flow_app/lib/config/routes/flow_routes.dart
@@ -1,9 +1,7 @@
-import 'package:college_flow_app/presentation/login/register_student.dart';
-
 abstract class FlowRoutes {
   static const mainPage = '/';
-  static const LoginPage = '/LoginPage';
-  static const RegisterPage = '/RegisterPage';
-  static const RegisterGeral = '/RegisterGeral';
-  static const RegisterStudent = '/RegisterStudent';
+  static const loginPage = '/LoginPage';
+  static const registerPage = '/RegisterPage';
+  static const registerGeral = '/RegisterGeral';
+  static const registerStudent = '/RegisterStudent';
 }
