+++ college_flow_app/lib/presentation/login/login_pages.dart
@@ -1,6 +1,6 @@
 import 'package:college_flow_app/config/design_system/data/colors/colors.dart';
 import 'package:college_flow_app/config/design_system/data/spacing/spacing.dart';
-import 'package:college_flow_app/presentation/login/register_page.dart';
+import 'package:college_flow_app/presentation/register/register_page.dart';
 import 'package:college_flow_app/presentation/widgets/buttons/flow_button.dart';
 import 'package:college_flow_app/presentation/widgets/gap.dart';
 import 'package:flutter/material.dart';
