+++ college_flow_app/lib/presentation/register/register_student.dart
@@ -2,27 +2,23 @@ import 'dart:ffi';
 
 import 'package:college_flow_app/config/design_system/data/colors/colors.dart';
 import 'package:college_flow_app/config/design_system/data/spacing/spacing.dart';
-import 'package:college_flow_app/presentation/login/register_geral.dart';
+import 'package:college_flow_app/presentation/register/register_adm.dart';
 import 'package:college_flow_app/presentation/widgets/buttons/flow_button.dart';
-import 'package:college_flow_app/presentation/widgets/checkboxlist.dart';
 import 'package:college_flow_app/presentation/widgets/gap.dart';
+import 'package:college_flow_app/presentation/widgets/input_text.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/src/foundation/key.dart';
 import 'package:flutter/src/widgets/framework.dart';
 
-class RegisterPage extends StatefulWidget {
-  const RegisterPage({Key? key}) : super(key: key);
+class RegisterStudent extends StatefulWidget {
+  const RegisterStudent({Key? key}) : super(key: key);
 
   @override
-  State<RegisterPage> createState() => _RegisterPageState();
+  State<RegisterStudent> createState() => _RegisterStudentState();
 }
 
-class _RegisterPageState extends State<RegisterPage> {
+class _RegisterStudentState extends State<RegisterStudent> {
   final _pageViewController = PageController();
-
-  bool _student = false;
-  bool _teacher = false;
-  bool _adm = false;
 
   @override
   Widget build(BuildContext context) {
@@ -60,52 +56,39 @@ class _RegisterPageState extends State<RegisterPage> {
             ),
             const VSpacer.xxs(),
             Text(
-              'Selecione seu tipo de usuário:',
-              style: Theme.of(context).textTheme.headlineSmall?.copyWith(
+              'Matrícula',
+              style: Theme.of(context).textTheme.titleMedium?.copyWith(
                     color: textColor,
                   ),
             ),
-            Image.asset('assets/icons/student_icon.png'),
             const VSpacer.nano(),
-            CheckboxList(
-              label: 'Aluno',
-              padding: const EdgeInsets.symmetric(horizontal: 118.0),
-              value: _student,
-              onChanged: (bool newValue) {
-                setState(() {
-                  _adm = false;
-                  _student = newValue;
-                  _teacher = false;
-                });
-              },
+            InputText(
+              placeholder: 'Digite sua Matrícula',
+              color: flowColorGrey,
             ),
-            Image.asset('assets/icons/professor_icon.png'),
-            const VSpacer.nano(),
-            CheckboxList(
-              label: 'Professor',
-              padding: const EdgeInsets.symmetric(horizontal: 108.0),
-              value: _teacher,
-              onChanged: (bool newValue) {
-                setState(() {
-                  _adm = false;
-                  _teacher = newValue;
-                  _student = false;
-                });
-              },
+            const VSpacer.xxxs(),
+            Text(
+              'Curso',
+              style: Theme.of(context).textTheme.titleMedium?.copyWith(
+                    color: textColor,
+                  ),
             ),
-            Image.asset('assets/icons/admin_icon.png'),
             const VSpacer.nano(),
-            CheckboxList(
-              label: 'Administrador',
-              padding: const EdgeInsets.symmetric(horizontal: 92.0),
-              value: _adm,
-              onChanged: (bool newValue) {
-                setState(() {
-                  _teacher = false;
-                  _adm = newValue;
-                  _student = false;
-                });
-              },
+            InputText(
+              placeholder: 'selecione seu Curso',
+              color: flowColorGrey,
+            ),
+            const VSpacer.xxxs(),
+            Text(
+              'Semestre',
+              style: Theme.of(context).textTheme.titleMedium?.copyWith(
+                    color: textColor,
+                  ),
+            ),
+            const VSpacer.nano(),
+            InputText(
+              placeholder: 'Digite seu Semestre',
+              color: flowColorGrey,
             ),
             const VSpacer.xxs(),
             FlowButton(
@@ -113,7 +96,7 @@ class _RegisterPageState extends State<RegisterPage> {
               onTap: () {
                 Navigator.of(context).push(
                   MaterialPageRoute(builder: (context) {
-                    return RegisterGeral();
+                    return RegisterAdministrator();
                   }),
                 );
               },
